delete unused code
