+++ packages/walt-compiler/src/core/string.js
@@ -1,23 +1,34 @@
-export default function stringPlugin() {
+export default function Strings() {
   return {
-    semantics() {
-      return {
-        StringLiteral: _ignore => args => {
-          const [stringLiteral, context] = args;
-          const { statics } = context;
-          const { value } = stringLiteral;
+    semantics: () => ({
+      CharacterLiteral: _ => ([node, context], transform) => {
+        const codePoint = node.value.codePointAt(0);
 
-          // did we already encode the static?
-          if (!(value in statics)) {
-            statics[value] = null;
-          }
+        return transform([
+          {
+            ...node,
+            Type: 'Constant',
+            type: 'i32',
+            value: String(codePoint),
+          },
+          context,
+        ]);
+      },
+      StringLiteral: _ignore => args => {
+        const [stringLiteral, context] = args;
+        const { statics } = context;
+        const { value } = stringLiteral;
 
-          // It's too early to tranform a string at this point
-          // we need additional information, only available in the generator.
-          // This also avoids doing the work in two places, in semantics AND gen
-          return stringLiteral;
-        },
-      };
-    },
+        // did we already encode the static?
+        if (!(value in statics)) {
+          statics[value] = null;
+        }
+
+        // It's too early to tranform a string at this point
+        // we need additional information, only available in the generator.
+        // This also avoids doing the work in two places, in semantics AND gen
+        return stringLiteral;
+      },
+    }),
   };
 }
